Split instance/new.rs and instance/convert.rs into separate files, add `ParametricInstance` struct.
